Update 00_austriaitaly.txt
I do not know if there is another event that triggers it sicne it's set to "triggered_only" I changed the script so it won't missfire but the changes are not applied (see "#") as I'm not familiar with the structure of the mod yet. If someone approves this edit, they should remove the # and change "triggered_only" to false. Thanks in advance :)
PS: I want to add a country flag so this event triggers only for that one specific war and not a conflict the two countries might end up to when fighting in a faction!