Some Americas stuff
Psychic spies from China try to steal your mind's elation
And little girls from Sweden dream of silver-screen quotation
And if you want these kind of dreams
It's Californication
It's the edge of the world, end all of western civilization
The Sun may rise in the East, at least it's settled in the final location
It's understood that Hollywood sells Californication
Pay your surgeon very well to break the spell of aging
Celebrity skin, is this your chin, or is that war you're waging?
Firstborn unicorn
Hardcore soft-porn
Dream of Californication
Dream of Californication
Dream of Californication
Dream of Californication
Marry me, girl, be my fairy to the world, be my very own constellation
A teenage bride with a baby inside, gettin' high on information
And buy me a star on the boulevard
It's Californication
Space may be the final frontier, but it's made in a Hollywood basement
And Cobain, can you hear the spheres, singing songs off station to station?
And Alderaan's not far away
It's Californication
Born and raised by those who praise control of population
Well, everybody's been there, and I don't mean on vacation
Firstborn unicorn
Hardcore soft-porn
Dream of Californication
Dream of Californication
Dream of Californication
Dream of Californication
Destruction leads to a very rough road, but it also breeds creation
And earthquakes are to a girl's guitar, they're just another good vibration
And tidal waves couldn't save the world from Californication
Pay your surgeon very well to break the spell of aging
Sicker than the rest, there is no test, but this is what you're craving
Firstborn unicorn
Hardcore soft-porn
Dream of Californication
Dream of Californication
Dream of Californication
Dream of Californication