Next Frenchnavyfocus
I have finished the next focus. The prerequisites for the decisions are still missing. All ships have the same equipment. The events are still without description. I made the focuses as I understood the text. So please look over it again to make sure everything is correct. The next focuses will follow in the next few days.